api/requirements.txt
aiosmtplib==1.1.7
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
black==22.8.0
blinker==1.5
celery==5.2.7
celerybeat-mongo==0.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
filelock==3.8.0
flower==1.2.0
greenlet==1.1.3
h11==0.13.0
humanize==4.4.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
loguru==0.6.0
Mako==1.2.3
MarkupSafe==2.1.1
mongoengine==0.24.2
mypy-extensions==0.4.3
nltk==3.8
numpy==1.23.5
packaging==21.3
pandas==1.5.3
passlib==1.7.4
pathspec==0.10.1
pipenv==2022.9.8
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
redis==4.3.4
regex==2022.10.31
requests==2.28.2
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.14
uvicorn==0.18.3
vine==5.0.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
